requirements.txt
attrs==25.4.0
beautifulsoup4==4.14.2
certifi==2025.11.12
cffi==2.0.0
charset-normalizer==3.4.4
h11==0.16.0
idna==3.11
numpy==2.3.4
outcome==1.3.0.post0
pandas==2.3.3
pyarrow==22.0.0
pycparser==2.23
PySocks==1.7.1
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
selenium==4.38.0
six==1.17.0
sniffio==1.3.1
sortedcontainers==2.4.0
soupsieve==2.8
trio==0.32.0
trio-websocket==0.12.2
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
websocket-client==1.9.0
wsproto==1.3.1
